feat: categories pages
