--- .env.exemple
@@ -1,1 +1,0 @@
-REACT_APP_API_URL=
