+++ src/components/App/App.jsx
@@ -10,16 +10,17 @@ import Category from '../Category/index.jsx';
 
 export default function App() {
     const [token, setToken] = useState('');
+    const [ categories, setCategories ] = useState([]);
 
     return (
-        <TokenContext.Provider value={{ token, setToken}}>
+        <TokenContext.Provider value={{ token, setToken, categories, setCategories}}>
             <BrowserRouter>
                 <Routes>
                     <Route path="/" element={<HomePage />} />
                     <Route path="/login" element={<LoginPage />} />
                     <Route path="/register" element={<CadastroPage />} />
                     <Route path="/user/shopping-cart" element={<ShoppingCart />} />
-                    <Route path="/category/:idCategory" element={<Category />} />
+                    <Route path="/category/:categoryName" element={<Category />} />
                 </Routes>
             </BrowserRouter>
         </TokenContext.Provider>
