+++ src/components/Header/index.jsx
@@ -1,33 +1,16 @@
-import { Category, Categories, Container, Header, MenuOptions, Title, UserOptions, Notification} from "./style.jsx";
+import { Container, MenuOptions, Notification, Title, UserOptions } from "./style";
 import { IoMdPerson, IoMdCart, IoMdMenu, IoMdClose } from "react-icons/io";
-import { Fragment, useContext,  useState, useEffect } from "react";
-import TokenContext from "../../contexts/TokenContext";
+import { useContext, useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
+import TokenContext from "../../contexts/TokenContext";
 import axios from "axios";
-import Book from "./../Book/index.jsx";
 
-export default function HomePage() {
+export default function Header() {
     const navigate = useNavigate();
-    const { token, setToken } = useContext(TokenContext);
+    const { token, setToken, categories, setCategories } = useContext(TokenContext);
     const [ menuUser, setMenuUser ] = useState(false);
     const [ menuOptions, setMenuOptions ] = useState(false);
-    const [ categories, setCategories ] = useState([]);
-    const [ booksData, setBooksData ] = useState([]);
     const [ shoppingCart, setShoppingCart ] = useState([]);
-
-    useEffect(() => {
-        const URL_CATEGORIES = `${process.env.REACT_APP_URL_API}/categories`;
-        const request = axios.get(URL_CATEGORIES);
-        request.then(response => setCategories(response.data.sort()))
-        request.catch(error => console.log(error));
-    },[]);
-
-    useEffect(() => {
-        const URL_BOOKS = `${process.env.REACT_APP_URL_API}/books`;
-        const request = axios.get(URL_BOOKS);
-        request.then(response => setBooksData(response.data))
-        request.catch(error => console.log(error));
-    },[]);
 
     useEffect(() => {
         if (token === "") return;
@@ -38,6 +21,14 @@ export default function HomePage() {
         request.then(response => setShoppingCart(response.data.shoppingCart))
         request.catch(error => console.log(error));
     },[token, shoppingCart]);
+
+
+    useEffect(() => {
+        const URL_CATEGORIES = `${process.env.REACT_APP_URL_API}/categories`;
+        const request = axios.get(URL_CATEGORIES);
+        request.then(response => setCategories(response.data.sort()))
+        request.catch(error => console.log(error));
+    },[setCategories]);
 
     function logout(){
         const URL_LOGOUT = `${process.env.REACT_APP_URL_API}/logout`;
@@ -52,14 +43,13 @@ export default function HomePage() {
 
     return (
         <Container>
-            <Header>
                 <section>
                     {(menuOptions===false) && <IoMdMenu size={20} color={"#00265d"}  
                     onClick={() => setMenuOptions(true)} />}
                     {(menuOptions===true) && <IoMdClose size={20} color={"#00265d"}  
                     onClick={() => setMenuOptions(false)} />}
                 </section>
-                <Title>BookStore</Title>
+                <Title onClick={() => navigate("/")}>BookStore</Title>
                 <section>
                     {(token!=="") && <IoMdCart size={20} color={"#00265d"}  
                     onClick={() => navigate("/user/shopping-cart")}/>} 
@@ -70,8 +60,7 @@ export default function HomePage() {
                     {(menuUser===true) && <IoMdClose size={20} color={"#00265d"}  
                     onClick={() => setMenuUser(false)}/>}
                 </section>
-            </Header>
-            {(token==="" && menuUser===true) &&
+                {(token==="" && menuUser===true) &&
                 <UserOptions>
                     <p onClick={() => navigate("/login")}>Entrar</p>
                     <p onClick={() => navigate("/register")}>Registre-se</p>
@@ -89,19 +78,7 @@ export default function HomePage() {
                     onClick={() => navigate(`/category/${categories}`)}>+ {categories}</p>)}
                 </MenuOptions>
             }
-            <Categories>
-                {categories?.map((category, i) => 
-                    <Fragment key = {i}>
-                        <h1>{category.toUpperCase()}</h1>
-                        <Category>
-                            {booksData?.filter(book => book.genero.includes(category)).map((book, key) => 
-                                <Book key={key} book={book}/>
-                            )}
-                        </Category>
-                    </Fragment>
-                    )}
-            </Categories>
-
         </Container>
-    )
+    );
+
 }
