+++ src/components/Login/index.jsx
@@ -3,23 +3,23 @@ import { useState,  useContext} from 'react';
 import { useNavigate } from 'react-router-dom';
 import TokenContext from '../../contexts/TokenContext';
 import loading from './../../assets/loading.gif';
-import { Button, Container, Form, Input, Registra, Title } from './style';
+import { Button, Container, Form, Input, Register, Title } from './style';
 
 export default function LoginPage(){
     const navigate = useNavigate();
     const { setToken } = useContext(TokenContext);
-    const [ dadosLogin, setDadosLogin ] = useState({email: "", hash: ""});
+    const [ loginData, setLoginData ] = useState({email: "", hash: ""});
     const [ disabled , setDisabled ] = useState(false);
 
     function login(e) {    
         e.preventDefault();
         const URL_LOGIN = `${process.env.REACT_APP_URL_API}/login`;
-        const promessa = axios.post(URL_LOGIN, dadosLogin);
+        const request = axios.post(URL_LOGIN, loginData);
         setDisabled(true);
-        promessa.then(resposta => {setToken(resposta.data);
+        request.then(resposta => {setToken(resposta.data);
                                     navigate("/");}
         );
-        promessa.catch(error => {setDisabled(false);
+        request.catch(error => {setDisabled(false);
                                 alert("Preencha os campos corretamente");}
         );
     }
@@ -28,19 +28,19 @@ export default function LoginPage(){
         <Container>
             <Title onClick={() => navigate("/")}>BookStore</Title>
             <Form onSubmit={login}>
-                <Input required type="email" placeholder=" E-mail" value={dadosLogin.email} 
-                onChange={e => setDadosLogin({...dadosLogin, email: e.target.value})} 
+                <Input required type="email" placeholder=" E-mail" value={loginData.email} 
+                onChange={e => setLoginData({...loginData, email: e.target.value})} 
                 pattern="^([\w\-]+\.)*[\w\- ]+@([\w\- ]+\.)+([\w\-]{2,3})$" 
                 title="Digite um endereço de email válido." disabled={disabled} />
-                <Input required type="password" placeholder="Senha" value={dadosLogin.hash} 
-                onChange={e => setDadosLogin({...dadosLogin, hash: e.target.value})} 
+                <Input required type="password" placeholder="Senha" value={loginData.hash} 
+                onChange={e => setLoginData({...loginData, hash: e.target.value})} 
                 pattern="^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[a-zA-Z]).{8,}$" disabled={disabled}
                 title="Mínimo de 8 caracterese pelo menos 1 letra maiúscula, 1 letra minúscula e 1 número" />
                 <Button type='submit' disabled={disabled}>
                     {disabled ? <img src={loading} alt="carregando..."/> : "Entrar"}
                 </Button>
             </Form>
-            <Registra to="/register" disabled={disabled}>Não tem uma conta? Registre-se!</Registra>
+            <Register to="/register" disabled={disabled}>Não tem uma conta? Registre-se!</Register>
             </Container>
     );
 }
