+++ src/index.js
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom';
 
-import App from './Components/App/App.jsx'
+import App from './components/App/App.jsx'
 import './reset/reset.css'
 
 ReactDOM.render(<App />, document.querySelector(".root"))
